+++ .dev_scripts/update_ut.py
@@ -15,10 +15,8 @@ parser.add_argument(
     default=[
         'mmedit/.mim', 'mmedit/registry.py', 'mmedit/version.py',
         '__pycache__', '__init__', '**/__init__.py', '**/stylegan3_ops/*',
-        '**/conv2d_gradfix.py', '**/grid_sample_gradfix.py',
-        '**/misc.py'
-        '**/upfirdn2d.py',
-        'mmedit/models/base_archs/all_gather_layer.py',
+        '**/conv2d_gradfix.py', '**/grid_sample_gradfix.py', '**/misc.py'
+        '**/upfirdn2d.py', 'mmedit/models/base_archs/all_gather_layer.py',
         '**/typing.py'
     ])
 args = parser.parse_args()
@@ -37,7 +35,7 @@ def update_ut():
     missing_ut = []
     blank_ut = []
 
-    file_list = glob(f'mmedit/**/*.py', recursive=True)
+    file_list = glob('mmedit/**/*.py', recursive=True)
 
     for f in tqdm(file_list):
         if check_exclude(f):
