+++ mmedit/models/__init__.py
@@ -2,8 +2,7 @@
 from ..registry import BACKBONES, COMPONENTS, LOSSES, MODELS
 from .base_models import (BaseConditionalGAN, BaseEditModel, BaseGAN,
                           BaseMattor, BaseTranslationModel, BasicInterpolator,
-                          ExponentialMovingAverage, MultiLayerDiscriminator,
-                          PatchDiscriminator)
+                          ExponentialMovingAverage)
 from .data_preprocessors import (EditDataPreprocessor, GenDataPreprocessor,
                                  MattorPreprocessor)
 from .editors import *  # noqa: F401, F403
@@ -11,8 +10,7 @@ from .losses import *  # noqa: F401, F403
 
 __all__ = [
     'BaseGAN', 'BaseTranslationModel', 'BaseEditModel', 'MattorPreprocessor',
-    'EditDataPreprocessor', 'BasicInterpolator', 'MultiLayerDiscriminator',
-    'BACKBONES', 'COMPONENTS', 'LOSSES', 'BaseMattor', 'MODELS',
-    'BasicInterpolator', 'ExponentialMovingAverage', 'GenDataPreprocessor',
-    'PatchDiscriminator', 'BaseConditionalGAN'
+    'EditDataPreprocessor', 'BasicInterpolator', 'BACKBONES', 'COMPONENTS',
+    'LOSSES', 'BaseMattor', 'MODELS', 'BasicInterpolator',
+    'ExponentialMovingAverage', 'GenDataPreprocessor', 'BaseConditionalGAN'
 ]
