+++ mmedit/models/base_models/__init__.py
@@ -1,7 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .average_model import ExponentialMovingAverage, RampUpEMA
+from .base_conditional_gan import BaseConditionalGAN
 from .base_edit_model import BaseEditModel
-from .base_gan import BaseConditionalGAN, BaseGAN
+from .base_gan import BaseGAN
 from .base_mattor import BaseMattor
 from .base_translation_model import BaseTranslationModel
 from .basic_interpolator import BasicInterpolator
