+++ mmedit/models/data_preprocessors/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .edit_data_preprocessor import EditDataPreprocessor, split_batch, stack_batch
+from .edit_data_preprocessor import (EditDataPreprocessor, split_batch,
+                                     stack_batch)
 from .gen_preprocessor import GenDataPreprocessor
 from .mattor_preprocessor import MattorPreprocessor
