+++ tests/test_models/test_base_archs/test_simple_encoder_decoder.py
@@ -2,7 +2,7 @@
 import numpy as np
 import torch
 
-from mmedit.models.base_models import SimpleEncoderDecoder
+from mmedit.models.base_archs import SimpleEncoderDecoder
 
 
 def assert_dict_keys_equal(dictionary, target_keys):
