+++ tests/test_models/test_base_archs/test_vgg.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 from mmengine.utils.dl_utils.parrots_wrapper import _BatchNorm
 
-from mmedit.models.base_models import VGG16
+from mmedit.models.base_archs import VGG16
 
 
 def check_norm_state(modules, train_state):
