+++ tests/test_models/test_editors/test_pix2pix/test_pix2pix.py
@@ -8,7 +8,8 @@ import torch
 from mmengine.logging import MessageHub
 from mmengine.optim import OptimWrapper, OptimWrapperDict
 
-from mmedit.models import GenDataPreprocessor, PatchDiscriminator, Pix2Pix
+from mmedit.models import GenDataPreprocessor, Pix2Pix
+from mmedit.models.base_archs import PatchDiscriminator
 from mmedit.models.editors.pix2pix import UnetGenerator
 from mmedit.utils import register_all_modules
