+++ tests/test_models/test_editors/test_stylegan2/test_ada/test_misc.py
@@ -1,0 +1,1 @@
+# Copyright (c) OpenMMLab. All rights reserved.
